+++ packages/gatsby/src/bootstrap/index.js
@@ -21,7 +21,7 @@ const loadThemes = require(`./load-themes`)
 const report = require(`gatsby-cli/lib/reporter`)
 import { getConfigFile } from "./get-config-file"
 const tracer = require(`opentracing`).globalTracer()
-const preferDefault = require(`./prefer-default`)
+import { preferDefault } from "./prefer-default"
 import { removeStaleJobs } from "./remove-stale-jobs"
 
 // Show stack trace on unhandled promises.
@@ -32,12 +32,12 @@ process.on(`unhandledRejection`, (reason, p) => {
 import { createGraphQLRunner } from "./create-graphql-runner"
 const { extractQueries } = require(`../query/query-watcher`)
 const requiresWriter = require(`./requires-writer`)
-const { writeRedirects } = require(`./redirects-writer`)
+import { writeRedirects, startRedirectListener } from "./redirects-writer"
 
 // Override console.log to add the source file + line number.
 // Useful for debugging if you lose a console.log somewhere.
 // Otherwise leave commented out.
-// require(`./log-line-function`)
+// import "./log-line-function"
 
 type BootstrapArgs = {
   directory: string,
@@ -69,6 +69,8 @@ module.exports = async (args: BootstrapArgs) => {
   // Start plugin runner which listens to the store
   // and invokes Gatsby API based on actions.
   startPluginRunner()
+
+  startRedirectListener()
 
   const directory = slash(args.directory)