+++ packages/gatsby-admin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.12...gatsby-admin@0.1.13) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-admin
+
 ## [0.1.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.11...gatsby-admin@0.1.12) (2020-05-05)
 
 **Note:** Version bump only for package gatsby-admin